new ui

+++ src/main/java/com/platonov/webapp/controller/HomeController.java
@@ -4,10 +4,17 @@ import com.platonov.webapp.domain.Status;
 import com.platonov.webapp.domain.User;
 import com.platonov.webapp.repo.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.List;
 
 @Controller
 @RequestMapping("/")
@@ -19,12 +26,8 @@ public class HomeController {
 
 
     @GetMapping
-    public String index(@AuthenticationPrincipal User user, Model model){
-        if(user !=null){
-            model.addAttribute("user",user.getUsername());
-        }
-        model.addAttribute("user","anonymous");
-        return "index";
+    public String index(@AuthenticationPrincipal User user){
+        return "redirect:/foruser";
     }
 
     @GetMapping("/login")
@@ -36,17 +39,34 @@ public class HomeController {
         model.addAttribute("users",userRepo.findAll());
         return "foruser";
     }
-    @GetMapping("user-delete/{id}")
-    public String deleteUser(@PathVariable("id") Long id){
-        userRepo.deleteById(id);
+    @RequestMapping(value = "/delete")
+    public String delete(@RequestParam("idChecked") Long[] ids){
+
+        if(ids != null)
+            for(Long id : ids)
+                userRepo.deleteById(id);
         return "redirect:/foruser";
     }
-    @GetMapping("user-block/{id}")
-    public String banUser(@PathVariable("id") Long id){
-        User user = userRepo.getById(id);
-        user.getStatus().clear();
-        user.getStatus().add(Status.BLOCK);
-        userRepo.save(user);
+    @RequestMapping(value = "/block")
+    public String blockUser(@RequestParam("idChecked") Long[] ids){
+        if(ids != null)
+            for(Long id : ids) {
+                User user = userRepo.getById(id);
+                user.getStatus().clear();
+                user.getStatus().add(Status.BLOCK);
+                userRepo.save(user);
+            }
+        return "redirect:/foruser";
+    }
+    @RequestMapping(value = "/unblock")
+    public String unblockUser(@RequestParam("idChecked") Long[] ids){
+        if(ids != null)
+            for(Long id : ids) {
+                User user = userRepo.getById(id);
+                user.getStatus().clear();
+                user.getStatus().add(Status.UNBLOCK);
+                userRepo.save(user);
+            }
         return "redirect:/foruser";
     }
     @GetMapping("user-unblock/{id}")
@@ -56,6 +76,14 @@ public class HomeController {
         user.getStatus().add(Status.UNBLOCK);
         userRepo.save(user);
         return "redirect:/foruser";
+    }
+    @RequestMapping(value="/logout", method = RequestMethod.GET)
+    public String logoutPage (HttpServletRequest request, HttpServletResponse response) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null){
+            new SecurityContextLogoutHandler().logout(request, response, auth);
+        }
+        return "redirect:/login?logout";
     }
 
 }
